LayoutTests/scrollingcoordinator/mac/rubberbanding-bounce-based-on-size-short-content-expected.txt

Swipes should not rubber-band on each side if the content size does not exceed the scroll view size.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Swipe to the left
PASS minXOffset is 0
PASS maxXOffset is -1000
PASS minYOffset is 0
PASS maxYOffset is -1000

Swipe to the right
PASS minXOffset is 0
PASS maxXOffset is -1000
PASS minYOffset is 0
PASS maxYOffset is -1000

Swipe down
PASS minXOffset is 0
PASS maxXOffset is -1000
PASS minYOffset is 0
PASS maxYOffset is -1000

Swipe up
PASS minXOffset is 0
PASS maxXOffset is -1000
PASS minYOffset is 0
PASS maxYOffset is -1000
PASS successfullyParsed is true

TEST COMPLETE
